Ditch the Splunk oneshot() API
The Splunk oneshot() API was super slow (seemingly CPU bound). Removed
it entirely and now use Splunk's Jobs API to retrieve results when the
'limit' paramter is specified. Further, we always use multiprocessing
in these cases, even if just with the default worker count of 1.
Limit searches are *much* faster now. Searches with no limit
specified still use the Splunk export() API, which is slower, but
ensures that all results are retrieved.