Initial version of OpenSearch compatibility.
This is a very janky compatibility, since it just restricts the Python "elasticsearch"
library version to anything before 7.14.0, when they added strict checking to make sure
it was connecting to an official Elastic cluster and not a fork (OpenSearch).

Also duplicated the existing Elastic unit tests with OpenSearch, so both are now tested.